% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.scalar2.R
\name{as.scalar2}
\alias{as.scalar2}
\title{opencpu thing to return value as a scalar instead of array.
NULL -> NA
not written by writer of package, but by someone at Roqua}
\usage{
as.scalar2(obj)
}
\arguments{
\item{obj}{Any object}
}
\value{
NA (if NULL) or obj as a scalar
}
\description{
opencpu thing to return value as a scalar instead of array.
NULL -> NA
not written by writer of package, but by someone at Roqua
}
