% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_graph.R
\name{comment_graph}
\alias{comment_graph}
\title{Create label with linebreaks from input-string for graphical purposes}
\usage{
comment_graph(input)
}
\arguments{
\item{input}{A string.}
}
\value{
A string of input potentially with linebreaks
}
\description{
Create label with linebreaks from input-string for graphical purposes
}
