% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json.R
\name{gettestfilepath}
\alias{gettestfilepath}
\title{get testfile path
tese functions are created by Roqua not by Gert Stulp, creator package}
\usage{
gettestfilepath(filename)
}
\arguments{
\item{filename}{the name of the file}
}
\value{
complete path
}
\description{
get testfile path
tese functions are created by Roqua not by Gert Stulp, creator package
}
