% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSON_to_DF.R
\name{JSON_to_DF}
\alias{JSON_to_DF}
\title{Create clean, long dataset from data received from Roqua}
\usage{
JSON_to_DF(raw)
}
\arguments{
\item{raw}{A dataframe in pre-specified format (429 variables)}
}
\value{
A clean dataframe that can be used for visualization
}
\description{
Create clean, long dataset from data received from Roqua
}
