% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Create clean, long dataset from data received from Roqua}
\usage{
prepare_data(data)
}
\arguments{
\item{data}{A dataframe in pre-specified format (429 variables)}
}
\value{
A clean dataframe that can be used for visualization
}
\description{
Create clean, long dataset from data received from Roqua
}
