% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_count.R
\name{time_count}
\alias{time_count}
\title{Create daynumber or weeknumber with weeks starting on Monday}
\usage{
time_count(dates, unit = "week")
}
\arguments{
\item{dates}{A vector with date-times.}

\item{unit}{A string giving time-unit ("week" or "day").}
}
\value{
A vector with day/weeknumbers.
}
\description{
Create daynumber or weeknumber with weeks starting on Monday
}
