% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_grid.R
\name{viz_grid}
\alias{viz_grid}
\title{Create ggplot-visualisation of 'grid' with experienced behaviours}
\usage{
viz_grid(data)
}
\arguments{
\item{data}{A dataframe in pre-specified format through "prepare_data"}
}
\value{
A ggplot-object ('grid'-visualisation)
}
\description{
Create ggplot-visualisation of 'grid' with experienced behaviours
}
