% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_nw.R
\name{viz_nw}
\alias{viz_nw}
\title{Create ggplot-visualisation of 'circle' with all symptoms}
\usage{
viz_nw(data, output = "poster")
}
\arguments{
\item{data}{A dataframe in pre-specified format through "prepare_data"}

\item{output}{String ("poster" or "slider") indicating which labels are presented}
}
\value{
A ggplot-object ('circle' visualisation)
}
\description{
Create ggplot-visualisation of 'circle' with all symptoms
}
