% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_poster.R
\name{viz_poster}
\alias{viz_poster}
\title{Create poster of all measurements}
\usage{
viz_poster(data, height = 60, width = 35)
}
\arguments{
\item{data}{A dataframe in pre-specified format through "prepare_data"}

\item{height}{height (dimension) of poster in XX}

\item{width}{width (dimension) of poster in XX}
}
\value{
An svg of poster
}
\description{
Create poster of all measurements
}
