% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_report_behaviour.R
\name{viz_report_behaviour}
\alias{viz_report_behaviour}
\title{Creates first visualisation for feedback-report (response-rate)}
\usage{
viz_report_behaviour(variable)
}
\arguments{
\item{data}{A dataframe with raw data}
}
\value{
An svg of visualisation
}
\description{
Creates first visualisation for feedback-report (response-rate)
}
