% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_report_response.R
\name{viz_report_response}
\alias{viz_report_response}
\title{Creates first visualisation for feedback-report (response-rate)}
\usage{
viz_report_response(variable, output_format = "svg")
}
\arguments{
\item{output_format}{String ("svg" or "ggplot") defining whether output should be ggplot or svg}

\item{data}{A dataframe with raw data}
}
\value{
An svg of visualisation
}
\description{
Creates first visualisation for feedback-report (response-rate)
}
