% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_slider.R
\name{viz_slider}
\alias{viz_slider}
\title{Create images for slides}
\usage{
viz_slider(data, height = 18.3, width = 36.8)
}
\arguments{
\item{data}{A dataframe in pre-specified format through "prepare_data"}

\item{height}{height (dimension) of image in cm}

\item{width}{width (dimension) of image in cm}
}
\value{
Multiple SVGs in string-format in list to be used in slider
}
\description{
Create images for slides
}
