% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_ts.R
\name{viz_ts}
\alias{viz_ts}
\title{Create ggplot-visualisation of lines over time (time-series)}
\usage{
viz_ts(data, left_right)
}
\arguments{
\item{data}{A dataframe in pre-specified format through "prepare_data"}

\item{left_right}{String ("left" or "right") indicating what sets of variables to plot}
}
\value{
A ggplot-object (time-series visualisation)
}
\description{
Create ggplot-visualisation of lines over time (time-series)
}
