% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_zoom.R
\name{viz_zoom}
\alias{viz_zoom}
\title{Creates see-through band for zooming on time-series graph of all measurements}
\usage{
viz_zoom(data_zoom)
}
\arguments{
\item{data}{A dataframe in pre-specified format through "prepare_data" only from period zoomed-in on}
}
\value{
A list with ggplot-layer that annotates graph with see-through "band" indicating period zoomed-in on
}
\description{
Creates see-through band for zooming on time-series graph of all measurements
}
