% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_report_alone.R
\name{viz_report_alone}
\alias{viz_report_alone}
\title{Creates first visualisation for feedback-report (alone by phase)}
\usage{
viz_report_alone(data)
}
\arguments{
\item{data}{A dataframe in pre-specified format through "prepare_data"}
}
\value{
An svg of visualisation
}
\description{
Creates first visualisation for feedback-report (alone by phase)
}
