% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_report_grid.R
\name{viz_report_grid}
\alias{viz_report_grid}
\title{Creates grid visualisation for feedback-report}
\usage{
viz_report_grid(data, output_format = "svg")
}
\arguments{
\item{data}{A dataframe in pre-specified format through "prepare_data"}

\item{output_format}{String ("svg" or "ggplot") defining whether output should be ggplot or svg}
}
\value{
A ggplot-object ('grid'-visualisation)
}
\description{
Creates grid visualisation for feedback-report
}
